update react related pacakges
