+++ .gitignore
@@ -27,3 +27,5 @@ yarn-error.log*
 
 # Created by github pipeline
 gha-creds-*.json
+
+.claude/
