+++ package.json
@@ -13,18 +13,18 @@
     "@reduxjs/toolkit": "^1.9.5",
     "@types/jest": "^27.0.1",
     "@types/node": "^16.7.13",
-    "@types/react": "^18.0.0",
-    "@types/react-dom": "^18.0.0",
+    "@types/react": "^19.2.14",
+    "@types/react-dom": "^19.2.3",
     "canvas-confetti": "^1.6.0",
     "dayjs": "^1.11.9",
     "fathom-client": "^3.5.0",
     "js-yaml": "^4.1.0",
     "lodash": "^4.17.21",
     "md5": "^2.3.0",
-    "react": "19.2.4",
-    "react-dom": "19.2.4",
-    "react-redux": "^8.0.5",
-    "react-router-dom": "7.12.0",
+    "react": "19.2.6",
+    "react-dom": "19.2.6",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "7.15.0",
     "react-timeago": "^7.2.0",
     "react-toastify": "11.0.5",
     "redux": "^4.2.1",
@@ -76,15 +76,15 @@
     "@types/uuid": "^9.0.8",
     "@typescript-eslint/eslint-plugin": "^6.1.0",
     "@typescript-eslint/parser": "^6.1.0",
-    "@vitejs/plugin-react": "^4.0.0",
+    "@vitejs/plugin-react": "^6.0.1",
     "eslint": "^8.42.0",
     "eslint-plugin-import-newlines": "^1.3.1",
     "eslint-plugin-newline-destructuring": "^1.2.1",
     "eslint-plugin-react": "^7.33.0",
     "prettier": "^2.8.8",
-    "vite": "^4.3.9",
-    "vite-plugin-checker": "^0.6.0",
-    "vite-plugin-svgr": "^3.2.0"
+    "vite": "^8.0.12",
+    "vite-plugin-checker": "^0.13.0",
+    "vite-plugin-svgr": "^5.2.0"
   },
   "resolutions": {
     "qs": "6.14.1"
